Removed "or Security Administrator"
Documentation states that you can be O365 GA or Security Admin, however, if you are not GA, you can't see the menu.
The below VSOs says that the documentation was going to be updated, but it appears that it was not.

Asking for the change based on the below information:

https://o365exchange.visualstudio.com/IP%20Engineering/_workitems/edit/768477
‘User with SCC Security Administrator permissions does not see "Attack Simulator" under Threat Management’
Ryan McGrath
commented Jun 12, 2018
Looks like this article has been corrected.  It used to say:
"You are an Office 365 global administrator, or you have the security administrator role assigned in the Security & Compliance Center."
It now says:
"You are an Office 365 global administrator assigned in the Security & Compliance Center"

Sunil Kadam
commented Jun 12, 2018
Attack simulator Is only available for Tenant administrators

https://o365exchange.visualstudio.com/IP%20Engineering/_workitems/edit/773156
‘Attack Simulator not accessable for Security Admins but is for Global Admins’
Stu Clark
commented May 22, 2018
Ok so that is a bug in the documentation, which is about to be refreshed.
AS when under preview supported the 2 roles, the documentation was still referring to Preview and had not been updated.
GA is currently Global Admin only, there is work to expand the roles back out again for other admin roles and conditional access.
Doc refresh should land this week.